fastapi/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
google-adk>=0.4
google-cloud-aiplatform>=1.70
vertexai>=1.70
pydantic>=2.0
python-dotenv>=1.0
